refactor: update file name and import to index.js
